Move NVCC compiler detection to the front
NVCC uses either gcc or clang depending on the host compiler being set
on the command line. Therefore it has to appear before both of them.
This makes predef detect nvcc as main compiler and gcc or clang as emulated.

The nvcc compilation process is somewhat special as can be read here:
https://docs.nvidia.com/cuda/cuda-compiler-driver-nvcc/index.html#cuda-compilation-trajectory
The nvcc compiler precompiles the input two times. Once for the device code
being compiled by the cicc device compiler and once for the host code
compiled by the real host compiler (gcc or clang).

So what predef (as preprocessor only lib) detects as compiler should be
the one doing the preprocessing which is NVCC in this case.

This behavior is similar to what boost config is doing in `select_compiler_config.hpp`.
There the NVCC detection is also the first. However, there it is not handled
as a real compiler (part of the #if-#elif) but as additional option before
the real compiler.

This is not possible in predef. Predef does not support two compilers being
set at the same time. So either NVCC or gcc/clang will be marked as emulated.

This change will possibly break code using predef with nvcc depending on
gcc or clang being said to be the main compiler.